Removed RoleApprovalRequest and transfered properties to ApprovalRequest
