+++ src/foam/nanos/approval/ApprovalRequest.js
@@ -18,20 +18,35 @@
   ],
 
   javaImports: [
-    'foam.core.FObject',
     'foam.core.X',
+    'foam.core.FObject',
+    'foam.dao.ArraySink',
     'foam.dao.DAO',
-    'foam.nanos.logger.Logger'
+    'foam.nanos.auth.*',
+    'foam.nanos.logger.Logger',
+    'foam.nanos.ruler.Operations',
+    'java.util.ArrayList',
+    'java.util.List',
+    'static foam.mlang.MLang.*'
+  ],
+
+  topics: [
+    'finished',
+    'throwError'
   ],
 
   requires: [
     'foam.dao.AbstractDAO',
-    'foam.nanos.approval.ApprovalStatus'
+    'foam.nanos.approval.ApprovalStatus',
+    'foam.u2.dialog.NotificationMessage'
   ],
 
   imports: [
     'approvalRequestDAO',
-    'ctrl'
+    'ctrl',
+    'currentMenu',
+    'stack',
+    'user'
   ],
 
   tableColumns: [
@@ -64,6 +79,52 @@
     }
   ],
 
+  axioms: [
+    {
+      class: 'foam.comics.v2.CannedQuery',
+      label: 'Pending',
+      predicateFactory: function(e) {
+        return e.EQ(
+          foam.nanos.approval.ApprovalRequest.STATUS,
+          foam.nanos.approval.ApprovalStatus.REQUESTED
+        );
+      }
+    },
+    {
+      class: 'foam.comics.v2.CannedQuery',
+      label: 'Approved',
+      predicateFactory: function(e) {
+        return  e.EQ(
+          foam.nanos.approval.ApprovalRequest.STATUS,
+          foam.nanos.approval.ApprovalStatus.APPROVED
+        );
+      }
+    },
+    {
+      class: 'foam.comics.v2.CannedQuery',
+      label: 'Rejected',
+      predicateFactory: function(e) {
+        return  e.EQ(
+          foam.nanos.approval.ApprovalRequest.STATUS,
+          foam.nanos.approval.ApprovalStatus.REJECTED
+        );
+      }
+    },
+    {
+      class: 'foam.comics.v2.CannedQuery',
+      label: 'All',
+      predicateFactory: function(e) {
+        return e.TRUE;
+      }
+    },
+    {
+      class: 'foam.comics.v2.namedViews.NamedViewCollection',
+      name: 'Table',
+      view: { class: 'foam.comics.v2.DAOBrowserView' },
+      icon: 'images/list-view.svg',
+    }
+  ],
+
   properties: [
     {
       class: 'Long',
@@ -78,25 +139,35 @@
       name: 'approver',
       section: 'requestDetails',
       documentation: `The user that is requested for approval. When set, "group" property is ignored.`,
-      tableCellFormatter: function(approver) {
+      view: function(_, X) {
+        if ( X.data.status === foam.nanos.approval.ApprovalStatus.REQUESTED ) {
+          return {
+            class: 'foam.u2.view.ValueView',
+            data$: X.data$.map((data) => data.REQUESTED)
+          };
+        } else {
+          return { class: 'foam.u2.view.ReferencePropertyView' };
+        }
+      },
+      tableCellFormatter: function(approver, data) {
         let self = this;
-        this.__subSubContext__.userDAO.find(approver).then((user)=> {
-          if ( user ) {
-            if ( self.__subSubContext__.user.id == user.id ) {
-              self.add(user.legalName);
-            } else {
-              self.add(user.group);
-            }
-          } else {
-            self.add(approver);
-          }
-        });
+        // If request is REQUESTED, show as Pending
+        // Otherwise, show approver's name
+        if ( data.status === foam.nanos.approval.ApprovalStatus.REQUESTED ) {
+          this.add(data.REQUESTED);
+        } else {
+          this.__subSubContext__.userDAO.find(approver).then(user => {
+            self.add(user ? user.toSummary() : `User #${approver}`);
+          });
+        }
       },
-      visibility: function(approver) {
-        return approver ?
-          foam.u2.DisplayMode.RO :
-          foam.u2.DisplayMode.HIDDEN;
-      },
+      visibility: function(status, approver) {
+        if ( status === foam.nanos.approval.ApprovalStatus.REQUESTED || ! approver ) {
+          return foam.u2.DisplayMode.HIDDEN;
+        }
+
+        return foam.u2.DisplayMode.RO;
+      }
     },
     {
       class: 'Object',
@@ -364,6 +435,56 @@
       class: 'Boolean',
       name: 'isTrackingRequest',
       value: false
+    },
+    {
+      class: 'Boolean',
+      name: 'isFulfilled',
+      visibility: 'HIDDEN'
+    },
+    {
+      class: 'Enum',
+      of: 'foam.nanos.ruler.Operations',
+      name: 'operation',
+      label: 'Action',
+      section: 'requestDetails',
+      visibility: function(operation) {
+        return operation ?
+          foam.u2.DisplayMode.RO :
+          foam.u2.DisplayMode.HIDDEN;
+      }
+    },
+    {
+      class: 'Reference',
+      of: 'foam.nanos.auth.User',
+      name: 'initiatingUser',
+      label: 'Requestor',
+      tableCellFormatter: function(initiatingUser) {
+        let self = this;
+        this.__subSubContext__.userDAO.find(initiatingUser).then(user => {
+          self.add(user ? user.toSummary() : `User #${initiatingUser}`);
+        });
+      },
+      section: 'requestDetails',
+      visibility: function(initiatingUser) {
+        return initiatingUser ?
+          foam.u2.DisplayMode.RO :
+          foam.u2.DisplayMode.HIDDEN;
+      }
+    }
+  ],
+
+  messages: [
+    {
+      name: 'SUCCESS_APPROVED',
+      message: 'You have successfully approved this request.'
+    },
+    {
+      name: 'SUCCESS_REJECTED',
+      message: 'You have successfully rejected this request.'
+    },
+    {
+      name: 'REQUESTED',
+      message: 'Pending'
     }
   ],
 
@@ -376,25 +497,34 @@
           type: 'Context'
         }
       ],
-      javaCode: `Logger logger = (Logger) x.get("logger");
-DAO dao = (DAO) x.get(getDaoKey());
-if ( dao == null ) {
-  logger.error(this.getClass().getSimpleName(), "DaoKey not found", getDaoKey());
-  throw new RuntimeException("Invalid dao key for the approval request object.");
-}
-FObject obj = dao.inX(x).find(getObjId());
-if ( obj == null ) {
-  logger.error(this.getClass().getSimpleName(), "ObjId not found", getObjId());
-  throw new RuntimeException("Invalid object id.");
-}
+      javaCode: `
+      Logger logger = (Logger) x.get("logger");
+      DAO dao = (DAO) x.get(getDaoKey());
+      if ( dao == null ) {
+        logger.error(this.getClass().getSimpleName(), "DaoKey not found", getDaoKey());
+        throw new RuntimeException("Invalid dao key for the approval request object.");
+      }
+
+      if ( getOperation() != Operations.CREATE ){
+        FObject obj = dao.inX(x).find(getObjId());
+        if ( obj == null ) {
+          logger.error(this.getClass().getSimpleName(), "ObjId not found", getObjId());
+          throw new RuntimeException("Invalid object id.");
+        }
+      }
       `
+    },
+    {
+      name: 'toSummary',
+      code: function() {
+        return `(${this.classification}) ${this.operation}`;
+      }
     }
   ],
 
   actions: [
     {
       name: 'approve',
-      label: 'Approve',
       section: 'requestDetails',
       isAvailable: (isTrackingRequest, status) => {
         if (
@@ -406,15 +536,30 @@ if ( obj == null ) {
         return ! isTrackingRequest;
       },
       code: function() {
-        this.status = this.ApprovalStatus.APPROVED;
-        this.approvalRequestDAO.put(this);
-        this.approvalRequestDAO.cmd(this.AbstractDAO.RESET_CMD);
-      },
-      tableWidth: 100
+        var approvedApprovalRequest = this.clone();
+        approvedApprovalRequest.status = this.ApprovalStatus.APPROVED;
+
+        this.approvalRequestDAO.put(approvedApprovalRequest).then(o => {
+          this.approvalRequestDAO.cmd(this.AbstractDAO.RESET_CMD);
+          this.finished.pub();
+          this.ctrl.add(this.NotificationMessage.create({
+            message: this.SUCCESS_APPROVED
+          }));
+
+          if ( this.currentMenu.id !== this.stack.top[2] ) {
+            this.stack.back();
+          }
+        }, e => {
+          this.throwError.pub(e);
+          this.ctrl.add(this.NotificationMessage.create({
+            message: e.message,
+            type: 'error'
+          }));
+        });
+      }
     },
     {
       name: 'reject',
-      label: 'Reject',
       section: 'requestDetails',
       isAvailable: (isTrackingRequest, status) => {
         if (
@@ -426,11 +571,27 @@ if ( obj == null ) {
         return ! isTrackingRequest;
       },
       code: function() {
-        this.status = this.ApprovalStatus.REJECTED;
-        this.approvalRequestDAO.put(this);
-        this.approvalRequestDAO.cmd(this.AbstractDAO.RESET_CMD);
-      },
-      tableWidth: 100
+        var rejectedApprovalRequest = this.clone();
+        rejectedApprovalRequest.status = this.ApprovalStatus.REJECTED;
+
+        this.approvalRequestDAO.put(rejectedApprovalRequest).then(o => {
+          this.approvalRequestDAO.cmd(this.AbstractDAO.RESET_CMD);
+          this.finished.pub();
+          this.ctrl.add(this.NotificationMessage.create({
+            message: this.SUCCESS_REJECTED
+          }));
+
+          if ( this.currentMenu.id !== this.stack.top[2] ) {
+            this.stack.back();
+          }
+        }, e => {
+          this.throwError.pub(e);
+          this.ctrl.add(this.NotificationMessage.create({
+            message: e.message,
+            type: 'error'
+          }));
+        });
+      }
     },
     {
       name: 'viewReference',
