+++ tools/classes.js
@@ -552,8 +552,7 @@ var classes = [
   'foam.nanos.approval.ApprovableApprovalRequestsPredicate',
   'foam.nanos.approval.ApprovableApprovalRequestsRule',
   'foam.nanos.approval.FulfilledApprovablePredicate',
-  'foam.nanos.approval.FulfilledApprovableRule',
-  'foam.nanos.approval.RoleApprovalRequest',
+  'foam.nanos.approval.FulfilledApprovableRule'
 ];
 
 var abstractClasses = [
